requirements.txt
pandas
numpy
matplotlib
transformers
torch
pip
pyarrow
nltk
scikit-learn
datasets
seqeval
banglanlptoolkit
bnunicodenormalizer
git+https://github.com/csebuetnlp/normalizer
langdetect
sentencepiece
protobuf
evaluate
accelerate 
python-dotenv